Add files via upload
Wrote two more functions called 'setMute' and 'previousSong'. The first will mute the volume of the current song upon a user's click interaction. The second will allow the user to move to the previous indexed song upon another click interaction or set the current progress of the song to zero if 3 seconds of the song has passed.