Add files via upload
Created the 'yourMusic.php' file which allows the user to create a new playlist. Once the 'new playlist' button is pressed, an alert prompts the user to input the name of the playlist, and then a MySQL query is initiated to enter the user's playlist name and a new playlist id into the local database.

In both the 'album.php' and 'artist.php' files, an on-click even was given to a newly created 'trackOptions' div. When a user clicks on the icon shown for each song, a menu will appear allowing the user to add or remove a song from a playlist.