Add files via upload
Wrote multiple functions to update the user's email and password. Both functions utilize an AJAX call to a separate PHP file to insert query into MySQL DB. There is also a new logout function that ends the user's session and redirects back to register.php. Additionally, there are three other functions, one that creates a playlist, another that deletes a playlist and one more that removes a song from a playlist. All three use an AJAX call to either insert or delete the corresponding data in the database.